pyproject.toml
[tool.poetry]
name = "eg-etl"
version = "0.0.1"
description = "EG ETL"
authors = ["Eric Yeung <ericox@gmail.com>"]
license = "proprietary"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
]
package-mode = false

[tool.poetry.dependencies]
python = "3.11.0"
jupyterlab = ">=4.0.5"
jupytext = ">=1.15.2"
pyspark = "3.4"
gocept-pseudonymize = "^3.0"
jupyter = "^1.0.0"
apache-airflow = "^2.8.2"
